Fix: reduce memory consumption
Instead of parsing every NASL plugin on load (can be up to 90k) it is
just parsing the includes and the root NASL plugin. With this change we
don't need to store a map with all information for a lookup table.

Currently tree-sitter is fast enough to not require threads for parallel
execution reducing the complexity of cache.rs significantly.