Refactor lookup and parser.rs is node_ext.rs
Move most functionality out of Lookup into into NASLInterpreter because I think
it is more fitting and changed Lookup to be a container for already handled
code.

Lookup is actually just needed for Block statements and mainly because the
parsing is written as Node extension which I'm trying to find a better solution
later solution int he hope that I can simplify it even further.